requirements.txt
fastapi[standard]
argon2-cffi
sqlmodel
psycopg2-binary